requirements.txt
django-cors-headers==3.3.0 # pip install django-cors-headers
djangorestframework==3.11.0 # pip install djangorestframework
vue-cli==2.9.6 # npm install -g vue-cli
bootstrap-vue==2.15.0 #npm i bootstrap-vue
nodejs==8.10.0# npm install nodejs
npm install vue-parallaxy@1.1.1
npm install --save axios@0.19.2 vue-axios

npm install
├── autoprefixer@7.2.6
├── axios@0.19.2
├── babel-core@6.26.3
├── babel-helper-vue-jsx-merge-props@2.0.3
├── babel-loader@7.1.5
├── babel-plugin-syntax-jsx@6.18.0
├── babel-plugin-transform-runtime@6.23.0
├── babel-plugin-transform-vue-jsx@3.7.0
├── babel-preset-env@1.7.0
├── babel-preset-stage-2@6.24.1
├── bootstrap-vue@2.15.0 extraneous
├── chalk@2.4.2
├── copy-webpack-plugin@4.6.0
├── css-loader@0.28.11
├── extract-text-webpack-plugin@3.0.2
├── file-loader@1.1.11
├── friendly-errors-webpack-plugin@1.7.0
├── html-webpack-plugin@2.30.1
├── node-notifier@5.4.3
├── nodejs@0.0.0 extraneous
├── optimize-css-assets-webpack-plugin@3.2.1
├── ora@1.4.0
├── portfinder@1.0.26
├── postcss-import@11.1.0
├── postcss-loader@2.1.6
├── postcss-url@7.3.2
├── rimraf@2.7.1
├── semver@5.7.1
├── shelljs@0.7.8
├── uglifyjs-webpack-plugin@1.3.0
├── url-loader@0.5.9
├── vue@2.6.11
├── vue-axios@2.1.5
├── vue-loader@13.7.3
├── vue-parallaxy@1.1.1 extraneous
├── vue-router@3.3.2
├── vue-style-loader@3.1.2
├── vue-template-compiler@2.6.11
├── webpack@3.12.0
├── webpack-bundle-analyzer@2.13.1
├── webpack-dev-server@2.11.5
└── webpack-merge@4.2.2

